feat: allow associated skill to list multiple skills (#1529)
This update revises how the Associated Skill field for items works. Now, the associated skill field can list multiple skills separated by the '|' (pipe) character (e.g. `Carousing|Zero-G|Demolitions`).  When the item is added to a character, the system will link that item to the highest level skill for that character based on the list.  Note that this linking only happens when adding (drag/dropping) the item to an actor. It does not dynamically update when skills change. Thanks to the suggestion of @comedian999.